back-helpdesk/app/core/config.py
import os
from datetime import timedelta
from dotenv import load_dotenv

# Carga el .env automáticamente
load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
if not DATABASE_URL:
    raise RuntimeError("DATABASE_URL no encontrado en .env")

JWT_SECRET = os.getenv("JWT_SECRET", "dev-secret")
JWT_ALG = "HS256"
ACCESS_TOKEN_EXPIRE = timedelta(hours=8)